Cost estimation should use the same price lookup logic as cost computing
